feat(fix): :sparkles: Colored Parity Sheep
In Minecraft Bedrock edition, the sheep get dyed on both the wool layer and the skin layer of the their texture. This feature brings consistency to the game by making the Minecraft Java edition sheep work the same way, instead of only having the wool texture dyed and the skin texture untouched.